Raised the photodiode sensor hardpoints.
I raised the photodiode sensor hardpoints so that they're level with the mic sensor hardpoint.  This should keep them from touching the ground.  I also decreased the diameter of the back sphere on which the robot rests from 9 to 8 mm so that the robot will be tilted a little further backwards.  Hopefully, this will make it more stable.